options.c: add DSCP code LE Least Effort
RFC-8622 implements a low priority DSCP marking called 'Least Effort' or
'LE'  Instead of prioritising traffic in varying degrees, this defines a
scum class of packet that really is the lowest of the low and you may
consider forwarding if you really have nothing better to do.

This patch adds LE class support to firewall3.

Signed-off-by: Kevin Darbyshire-Bryant <ldir@darbyshire-bryant.me.uk>